requirements_ml.txt
# requirements_ml.txt - Enhanced requirements for ML-powered MindSpace

# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==3.0.1

# OpenAI Integration
openai==1.3.7

# Machine Learning and Data Science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2

# Natural Language Processing
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0

# Additional ML utilities
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Text preprocessing and analysis
nltk==3.8.1
textblob==0.17.1
wordcloud==1.9.3

# Progress bars and utilities
tqdm==4.66.1

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

# Data validation and processing
marshmallow==3.20.1
cerberus==1.3.5

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Security enhancements
cryptography==41.0.8
bcrypt==4.1.2

# Optional: Database migrations
Flask-Migrate==4.0.5

# Optional: API documentation
flasgger==0.9.7.1

# Optional: Caching
Flask-Caching==2.1.0
redis==5.0.1

# Optional: Task queue for background processing
celery==5.3.4
kombu==5.3.4

# Optional: For advanced features
# plotly==5.15.0  # Interactive charts
# streamlit==1.25.0  # Alternative UI
# fastapi==0.103.0  # API endpoints
# uvicorn==0.23.2  # ASGI server